+++ packages/winnow/CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [1.16.10] - 11-05-2019
 ### Fixed
 * Fix SQL generated when using group by with where clause.
 * Split comma delimited group by fields
@@ -317,6 +317,7 @@ Deep clone for dates since mutating dates has down stream consequences if reused
 ### Added
 * Initial release
 
+[1.16.10]: https://github.com/featureserver/winnow/compare/v1.16.9...v1.16.10
 [1.16.9]: https://github.com/featureserver/winnow/compare/v1.16.8...v1.16.9
 [1.16.8]: https://github.com/featureserver/winnow/compare/v1.16.7...v1.16.8
 [1.16.7]: https://github.com/featureserver/winnow/compare/v1.16.6...v1.16.7